refactor lambda
