+++ Lambda/translate/__init__.py
@@ -1,0 +1,3 @@
+from flask import Flask, render_template
+
+app = Flask(__name__)
