+++ Lambda/translate/__pycache__/__init__.cpython-310.pyc
