+++ Lambda/translate/__pycache__/run.cpython-310.pyc
