+++ Lambda/translate/handler.py
@@ -1,0 +1,4 @@
+from run import app
+
+def lambda_handler(event, context):
+    app.run(debug=True)
