+++ Lambda/translate/library/db_languages.py
@@ -1,0 +1,12 @@
+global_lang = [
+    {
+        'name': 'English',
+        'code': 'en'
+    },
+    {
+        'name': 'French',
+        'code': 'fr'
+    }
+]
+
+local_lang = ['Kom', 'Bafut']
