+++ Lambda/translate/library/english_kom.csv
@@ -1,0 +1,117 @@
+english,kom
+mouth,ɨkfvɨ
+eye,Isɨ
+head,Atú
+hair,Nyungsɨ Atu
+tooth,IsÔng
+tongue,Ilemí
+nose,Ayui
+ear,Atunglɨ
+neck,Asɨɨtong
+back_of_neck,Ibam itong
+throat,Imzî
+breast,Igheyn
+arm/hand,ɨkue
+claw,Ngaesɨ
+nail,Ngae Awu
+leg,ɨfieyn
+foot,Avɨ
+buttock,ɨbuinasas
+belly,ɨlva
+navel,Itong
+intestines,Mɨtú
+insides,Mɨtú
+blood,Mɨlung
+urine,Mindzui
+bone,Avɨf
+skin,Iwuyn
+wing,IwÔ
+feather,Tɨwho
+horn,Ndong
+tail,ɨnkuin
+human being,Wúl
+man,Wulɨlumnɨ
+male,Wulɨlumnɨ
+woman,Wulɨwi
+husband,Lúm
+child,Wȃyn
+name,Iyɨl
+sky,Ivɨ
+night,Nintu’ɨ
+moon,Njong
+sun,Chuí
+wind,Afɨef
+cloud,Mbahsɨ
+dew,Amua’
+rain,Ivɨ
+ground,Nse
+sand,Asese
+path,Dzɨ
+water,Mû
+stream,Gvȃ
+river,Gvȃ
+house,Ndo
+fire,ɨvɨs
+firewood,ɨnka’
+smoke,Fɨle’
+ash,Abzitɨ Ichɨ
+animal,Nyȃm
+meat,Njɨuo
+dog,Bɨ́
+elephant,Atȃm
+goat,Bzí
+bird,Fɨnyûyn
+tortoise,Tolaki
+snake,Mbam
+fish,Fvî
+one,No’
+two,Bo
+three,Tal
+four,Kae
+five,Tayn
+six,Ntufa
+seven,SÔmbo
+eight,Nfama
+nine,Bɨlamo’
+ten,Ivɨm
+come,Gvi
+send,Túm
+walk,Jel
+fall,Fè
+leave,Lu-a
+fly,Tayn
+pour,Bungsɨ
+strike,Chiɨ’sɨ
+bite,Lúm
+wash,Sû
+split,Sȃyn
+give,Fú
+steal,Chóng
+squeeze,Kám
+cultivate,Fèl
+bury,Gvɨmi
+burn,Nyo’
+eat,Yɨ
+drink,Nyvɨ
+vomit,Ifɨsɨ
+suck,Nyong
+spit (saliva),Minchɨa
+saliva,Minchɨa
+blow,Fɨef
+on,Fɨef
+swell,Mùl
+give birth,Bzi
+die,Kfɨ
+kill,Zui
+push,Tɨyn
+pull,Guf
+sing,Yêm
+play,Dɨm
+be afraid,Ifayn
+want,Kɨng
+say,Bae
+see,Yeyn
+show,Dɨen
+hear,Yvɨ
+know,Kɨa
+count,Táng
