+++ Lambda/translate/manage_globals/__pycache__/main.cpython-310.pyc
