+++ Lambda/translate/requirements.txt
