+++ Lambda/translate/templates/index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>JSC Translator</title>
+</head>
+
+<body>
+  <h1>Custom Language Translator</h1>
+  <form action="/translate" method="post">
+    <label for="input_text">Enter Text:</label>
+    <textarea name="input_text" id="input_text" rows="4" cols="50" required></textarea>
+    <br>
+    <button type="submit">Translate</button>
+  </form>
+  {% if input_text %}
+  <p>Input Text: {{ input_text }}</p>
+  {% endif %}
+  {% if translated_text %}
+  <p>Translated Text: {{ translated_text }}</p>
+  {% endif %}
+</body>
+
+</html>
