--- application/__init__.py
@@ -1,3 +1,0 @@
-from flask import Flask, render_template
-
-app = Flask(__name__)
