--- application/library/db_languages.py
@@ -1,12 +1,0 @@
-global_lang = [
-    {
-        'name': 'English',
-        'code': 'en'
-    },
-    {
-        'name': 'French',
-        'code': 'fr'
-    }
-]
-
-local_lang = ['Kom', 'Bafut']
