--- run_tests.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-pytest --rootdir=/home/runner/work/app-jsc-translator-be/application/tests/test_translator.py
